Reorganized class.c, class_prefabs.c, and class_skills.c: - class.c:   - moved class_init_skills() (or class_skills_init(), I forget the name) to class_prefabs.   - added in its place class_add_skills(), which takes a precreated skilltree and skill inventories and adds them to a class.   - Restyled the file to my liking. - class_skills.c:   - The main function there became class_allocate_skills() in class_prefabs.c   - The secondary function also moved to class_prefabs_add_skills. - class_prefabs:   - Received afformentioned functions - Also modified .h files, Cmake files, and a test case. - WHY:   - The functions moved to class_prefab only worked for prefab classes.  Custom classes needs a different way to add its own     external trees and skill inventories, which was added.   - In fact, the custom class function might end up being used by the prefab one!
