+++ src/playerclass/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(playerclass
         src/class.c
         src/class_item_restriction.c
         src/class_prefabs.c
-        src/class_skills.c
         )
 
 target_include_directories(playerclass PRIVATE src/)
