+++ src/playerclass/src/class.c
@@ -11,7 +11,7 @@
 
 #include "playerclass/class_structs.h"
 #include "playerclass/class.h"
-#include "class_skills.h"
+#include "playerclass/class_prefabs.h"
 #include "common/utlist.h"
 
 /* Rudimentary id system for prefab classes (internal) */
@@ -63,22 +63,20 @@ int get_class_name_index(char* name) {
 
 /* See class.h */
 class_t* class_new(char* name, char* shortdesc, char* longdesc,
-                   obj_t* attr, stats_hash_t* stat, effects_hash_t* effect)
-{
+                   obj_t* attr, stats_hash_t* stat, effects_hash_t* effect) {
     int rc;
     class_t* c;
 
     c = (class_t*) calloc(1, sizeof(class_t));
 
-    if(c == NULL)
-    {
+    if(c == NULL) {
         fprintf(stderr, "Could not allocate space for class_new.\n");
         return NULL;
     }
 
     rc = class_init(c, name, shortdesc, longdesc, attr, stat, effect);
 
-    if (rc == EXIT_FAILURE){
+    if (rc == EXIT_FAILURE) {
         fprintf(stderr, "Could not initalize values for class_new.\n");
     }
 
@@ -87,25 +85,21 @@ class_t* class_new(char* name, char* shortdesc, char* longdesc,
 
 /* See class.h */
 int class_init(class_t* class, char* name, char* shortdesc, char* longdesc,
-               obj_t* attr, stats_hash_t* stat, effects_hash_t* effect)
-{
-    if (class == NULL)
-    {
+               obj_t* attr, stats_hash_t* stat, effects_hash_t* effect) {
+    if (class == NULL) {
         fprintf(stderr, "Class to initialize in class_init is NULL.\n");
         return EXIT_FAILURE;
     }
 
     class->name = (char*) calloc(MAX_NAME_LEN + 1, sizeof(char));
-    if (class->name == NULL)
-    {
+    if (class->name == NULL) {
         fprintf(stderr, "Could not allocate memory for name in class_init.\n");
         return EXIT_FAILURE;
     }
     strncpy(class->name, name, MAX_NAME_LEN);
 
     class->shortdesc = (char*) calloc(MAX_SHORT_DESC_LEN + 1, sizeof(char));
-    if (class->name == NULL)
-    {
+    if (class->name == NULL) {
         fprintf(stderr, "Could not allocate memory for short description in "
                         "class_init.\n");
         return EXIT_FAILURE;
@@ -113,8 +107,7 @@ int class_init(class_t* class, char* name, char* shortdesc, char* longdesc,
     strncpy(class->shortdesc, shortdesc, MAX_SHORT_DESC_LEN);
 
     class->longdesc = (char*) calloc(MAX_LONG_DESC_LEN + 1, sizeof(char));
-    if (class->name == NULL)
-    {
+    if (class->name == NULL) {
         fprintf(stderr, "Could not allocate memory for long description in "
                         "class_init.\n");
         return EXIT_FAILURE;
@@ -134,75 +127,36 @@ int class_init(class_t* class, char* name, char* shortdesc, char* longdesc,
 }
 
 /* See class.h */
-int class_init_skills(class_t* class, int max_skills_in_tree, 
-                      int max_active_skills, int max_passive_skills) 
-{
-    class->combat = inventory_new(max_active_skills, max_passive_skills);
-    class->noncombat = inventory_new(max_active_skills, max_passive_skills);
-    
-    /* tree ID needs to be unique across all chiventure code.  Our team has been
-     * assigned the range 3000-3999.  Default classes start at 3000. There is
-     * currently no support for non-prefab classes. */
-    int class_id = get_class_name_index(class->name);
-    if (class_id == -1)
-    {
-        fprintf(stderr, "Could not find class name: \"%s\" "
-                        "in class_init\n", class->name);
-        return EXIT_FAILURE;
-    }
-    int tid = 3000 + class_id;
-    class->skilltree = skill_tree_new(tid, class->name, max_skills_in_tree);
-
-    if (class->skilltree == NULL)
-    {
-        fprintf(stderr, "Could not allocate memory for skill trees "
-                        "in class_init\n");
-        return EXIT_FAILURE;
-    }
-    if (class->combat == NULL || class->noncombat == NULL)
-    {
-        fprintf(stderr, "Could not allocate memory for skill inventories"
-                        "in class_init\n");
-        return EXIT_FAILURE;
-    }
-    if (class_add_skills(class) == EXIT_FAILURE)
-    {
-        return EXIT_FAILURE;
-    }
-
-    return EXIT_SUCCESS;
+int class_add_skills(class_t* class, skill_inventory_t* combat, 
+                     skill_inventory_t *noncombat, skill_tree_t* skilltree) {
+    class->combat = combat;
+    class->noncombat = noncombat;
+    class->skilltree = skilltree;
+    return SUCCESS;
 }
 
 /* See class.h */
-int class_free(class_t* class)
-{
-    if (class == NULL)
-    {
+int class_free(class_t* class) {
+    if (class == NULL) {
         return EXIT_SUCCESS;
     }
 
-    if (class->name != NULL)
-    {
+    if (class->name != NULL) {
         free(class->name);
     }
-    if (class->shortdesc != NULL)
-    {
+    if (class->shortdesc != NULL) {
         free(class->shortdesc);
     }
-    if (class->longdesc != NULL)
-    {
+    if (class->longdesc != NULL) {
         free(class->longdesc);
     }
-    if (class->skilltree != NULL)
-    {
+    if (class->skilltree != NULL) {
         skill_tree_free(class->skilltree);
     }
-    if (class->combat != NULL)
-    {
+    if (class->combat != NULL) {
         inventory_free(class->combat);
     }
-    if (class->noncombat != NULL)
-    {
+    if (class->noncombat != NULL) {
         inventory_free(class->noncombat);
     }
